v2_dev/requirements.txt
streamlit
azure-ai-formrecognizer
azure-core
python-dotenv
factur-x
lxml
pypdf
pandas
openpyxl